Significant re-working of internal printing
This commit shifts the determination of 'ansi' or 'universal' printing
modes to run time. There is now a subset,
Terminal::Print::MoveCursorProfile, that is utilized for safeguarding
the constructors for both Terminal::Print and Terminal::Print::Grid. We
enable 'ansi' by default. This  subset will be exported into any
namespace which imports Terminal::Print::Commands.

Thanks to nine++ for pointing towards a better API that allows the
Terminal::Print user to choose a profile at run time. A program could
offer a debug or setup mode that would prompt the user with a test
movement and then ask or confirmation that it worked. If not, re-create
as 'universal'.

Meanwhile, I promised up the .change-cell and .print-cell methods in a
way that should a) remain transparent at the API level, and b) seems to
me to actually be quite a bit smoother/faster at printing. I'm curious
to hear whether this might be the case for others.