Buff up the SIGINT handler
If we die in the sigint handler we can be rest assured that all the relevant
supplies will be chopped off and we will really exit at the time that
the user so explicitly requested.

This also moves the $!print-enabled pattern from the Terminal::Print
object to its grids. This should avoid race conditions (due to
OO::Monitors) as well as keeping it a bit cleaner in my opinion.

Later, we can harden these cases so that you can't double disable a grid,
maybe throw an exception when trying to print to a disabled grid.