Avoid 'no precompilation' infection from T::P::Dimensions
'no precompilation' is transitive, so a module that uses a 'no precompilation'
module cannot be precompiled either.  Avoid this infection by dropping use
of T::P::Dimensions (folding in the columns and rows subs) and doing tput
calls within an INIT phaser rather than a BEGIN phaser.

This is just as dynamic as before (and clearer in intent), and calls tput
just as often at every process start, but is faster as it allows the *code*
of all modules to be precompiled again.  (How much faster will depend on how
many modules transitively depend on T::P::Commands in your program.)