Remove noop coercion
The &( ) coercer doesn't actually do anything (except possibly itemize the
interior expression), and even if it did do something having to do with
Callable (the Role for the '&' sigil), that would be the wrong thing to do
here.

In other words, this only ever worked because Rakudo silently ignored it.