First sketch of RPG UI simulation
In order to get a good idea of real-world requirements and challenges,
create a simulated CRPG or Roguelike UI and run it through a few common
operations.

This commit is a very first sketch, using only ASCII, no real colors, and
with only the barest outline of the functionality to come.  On the flip side,
it does offer a --bench switch to turn on benchmark mode (which turns off
sleeps and rate limits, trying to complete all operations as fast as possible).