More Widget refactoring in rpg-ui example
Give every Widget a required height and width, and make ProgressBar fill that
space; set the loading screen ProgressBar to be height 3 to confirm this works.
(I'm ignoring the problem of duplicated ProgressBar text for now.)

Unfortunately, it also confirms that drawing the ProgressBar is slow and
something needs to be done about the visual tearing.