Add a KeyframeAnimation widget to rpg-ui
Takes a series of keyframe grids, and animates between them using various
effects.  For now, only one effect method is available, 'speckle'.  As each
keyframe is reached, the current keyframe number is emitted to its on-keyframe
Supplier attribute.  Before starting the animation, you can .Supply.tap this
Supplier to follow the progress of the animation occuring in its background
thread.