Work around precomp issues, part 1
Improve handling of Terminal::Print::Commands::{columns,rows} to avoid strange
bug when precompiling that would freeze the detected terminal size at the size
that the precomp code uses, which is 80x24, *no matter how big* the actual
terminal may be.  This shouldn't happen at all, because terminal size should
be detected at INIT time, not BEGIN time (it's right there in the code!) but
apparently that's not what actually happens.