A couple of improvements to Terminal::Print
Allow Terminal::Print to be used wrapped by another framework and
provide support for resizing:

- Replace initialization in method `new` with a dedicated `setup` method
invoked from `TWEAK`.

- `setup` method supports `:reset` adverb causes `Terminal::Print`
object to be reconfigured. This allows to resize it on the fly.

- A grid created by default get `.default` name. This is the grid to be
re-configured by `setup(:reset)`

- It is possible to create `root-widget` of any custom type inheriting
from `Terminal::Print::Widget` as well as provide pre-created object to
act as the root one.

- Methods `grid` and `grid-object` check for existence of a named grid
in `%!grid-name-map`

Minor fix for `grid-object` proto.