Fix braino that made T.blit/print-grid do nothing
Line 251 of Print.pm6 used to be this line:

    proto method print-grid($) { }

It took disturbingly long for me to realize that should have been:

    proto method print-grid($) { * }

Because of the missing '*', the proto was not an OnlyStar, which just passes
on the call to the multis below.  Instead it was an empty routine, which just
immediately returned.  In other words, T.print-grid (and therefore T.blit)
*never actually worked*!

Raku doesn't actually need such a simple proto, so instead I decided to just
delete it instead.

Fixes #73.